Plot on vsync, not on packet receive (#138)
Update the plot on vsync. Results in smoother plotting on high refresh
rate screens.

+++ test/segmented_graph_cache_test.dart
@@ -22,8 +22,8 @@ class _Harness {
   /// fake renderer reports tail provisionality the same way the envelope
   /// renderer does (bake end within one tail span of the data edge).
   bool paint({
-    required int viewStart,
-    required int viewSpan,
+    required double viewStart,
+    required double viewSpan,
     required int totalSamples,
     double gw = 400,
     double gh = 100,
